iwinfo: use upstream functions for channel-frequency mapping
the kernel and iw utility use the following functions
to convert channel to frequency and frequency to channel
because the channel numbers and frequencies can overlap.
they are updated and more accurate.

make them generic, not relying on nl80211 api macros
and put them in utils.c so they can be used anywhere

new function mode2band() is needed
to map the "hw_mode" value from hostapd
to the matching int representing the band
in the enum nl80211_band

to handle 6 GHz band with hostapd, which shares the same
"hw_mode" value as the 5 GHz band, look for the option
"op_class" which we only use for 6 GHz.

Signed-off-by: Michael Pratt <mcpratt@pm.me>