iwinfo: handle truncate compilation warning
recent versions of GCC warn when the max length
of a string written to a buffer with snprintf()
can be larger than the size of that buffer.

this warning can be handled by either
increasing buffer sizes, or,
checking the return value of snprintf()

bump the buffer for ifname path to 128
to match the others.

Signed-off-by: Michael Pratt <mcpratt@pm.me>